+++ asafov.andrew/P3/matrixAutoInitiallization.hpp
@@ -2,6 +2,6 @@
 #define MATRIXAUTOINITIALLIZATION_H
 namespace asafov
 {
-  void matrixAutoInitialization(int* mtx, char* input, int& n, int& m);
+  void matrixAutoInitialization(int* mtx, char* input, size_t& n, size_t& m);
 }
 #endif
