+++ asafov.andrew/P3/matrixDynamicInitiallization.hpp
@@ -2,6 +2,6 @@
 #define MATRIXDYNAMICINITIALLIZATION_H
 namespace asafov
 {
-  void matrixDynamicInitialization(int* mtx, char* input, int& n, int& m);
+  void matrixDynamicInitialization(int* mtx, char* input, size_t& n, size_t& m);
 }
 #endif
