+++ asafov.andrew/P3/matrixHandling.hpp
@@ -2,7 +2,7 @@
 #define MATRIXHANDLING_H
 namespace asafov
 {
-  void matrixHandling(int* mtx, int n, int m);
-  int matrixHandlingSanctions(int* mtx, int n, int m);
+  void matrixHandling(int* mtx, size_t n, size_t m);
+  int matrixHandlingSanctions(int* mtx, size_t n, size_t m);
 }
 #endif
